# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Cards REST API
  description: API для управления пользователями и банковскими картами
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local server
paths:
  /auth/sign-in:
    post:
      summary: Аутентификация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCredentialsDto'
      responses:
        '200':
          description: JWT токен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtAuthenticationDto'
  /auth/refresh:
    post:
      summary: Обновление JWT токена
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenDto'
      responses:
        '200':
          description: Новый JWT токен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtAuthenticationDto'
...
components:
  schemas:
    UserCredentialsDto:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    JwtAuthenticationDto:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
    RefreshTokenDto:
      type: object
      properties:
        refreshToken:
          type: string
...